# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  commonerrors.ErrorMap:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  oauthDTO.Response:
    properties:
      is_active:
        type: boolean
      message:
        type: string
      oauth_token:
        type: string
      oauth_url:
        type: string
      status:
        type: string
    type: object
  userDTO.GenericResponse:
    properties:
      data: {}
      error:
        type: string
      status:
        type: string
    type: object
...
  userDTO.ResetPassword:
    properties:
      new_password1:
        type: string
      new_password2:
        type: string
      otp:
        type: string
    required:
    - new_password1
    - new_password2
    - otp
    type: object
...
info:
  contact: {}
  description: API for Rasta
  title: Rasta API
  version: "1.0"
paths:
...
  /users/oauth/disable:
    delete:
      consumes:
      - application/json
      description: Disables OAuth for the user's account by verifying the provided
        OAuth code.
      parameters:
      - description: OAuth code
        in: body
        name: oauth
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OAuth disabled successfully
          schema:
            $ref: '#/definitions/oauthDTO.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      security:
      - BearerAuth: []
      summary: Disable OAuth
      tags:
      - OAuth
  /users/oauth/enable:
    post:
      consumes:
      - application/json
      description: Verifies the OAuth code provided by the user and enables OAuth
        for the account.
      parameters:
      - description: OAuth code
        in: body
        name: oauth
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OAuth enabled successfully
          schema:
            $ref: '#/definitions/oauthDTO.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      security:
      - BearerAuth: []
      summary: Verify and Enable OAuth
      tags:
      - OAuth
  /users/oauth/generate:
    get:
      description: Generates an OAuth secret and URL for the user to enable OAuth.
      produces:
      - application/json
      responses:
        "200":
          description: OAuth secret and URL
          schema:
            $ref: '#/definitions/oauthDTO.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      security:
      - BearerAuth: []
      summary: Generate OAuth Secret and URL
      tags:
      - OAuth
  /users/otp/{id}/verify:
    post:
      consumes:
      - application/json
      description: Verifies the user's email using the provided OTP. If successful,
        marks the email as verified and deletes the OTP.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: OTP code
        in: body
        name: otp
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      summary: Verify Email with OTP
      tags:
      - OTP
  /users/otp/resend:
    post:
      consumes:
      - application/json
      description: Resends the OTP to the user's email for verification purposes.
      parameters:
      - description: User email
        in: body
        name: email
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      summary: Resend OTP to Email
      tags:
      - OTP
  /users/reset-password:
    get:
      consumes:
      - application/json
      description: Generates a password reset code and sends it to the user's email.
      parameters:
      - description: User email
        in: body
        name: email
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password reset code sent successfully
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      summary: Send Password Reset Code
      tags:
      - Password Reset
  /users/reset-password/{id}/verify:
    post:
      consumes:
      - application/json
      description: Verifies the provided OTP and, if valid, allows the user to reset
        their password.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Password reset request body
        in: body
        name: ResetPassword
        required: true
        schema:
          $ref: '#/definitions/userDTO.ResetPassword'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            $ref: '#/definitions/userDTO.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      summary: Verify OTP and Reset Password
      tags:
      - Password Reset
...
swagger: "2.0"